refactor: organize API endpoints and add Swagger UI
